Merge Upstream changes

+++ CO2-Ampel_Plus/MQTTManager.h
@@ -7,7 +7,7 @@
 #include <ArduinoJson.h>
 
 bool mqtt_connect();
-void mqtt_send_value(int co2, int temp, int hum, int lux);
+void mqtt_send_value(int co2, float temp, int hum, int lux);
 bool mqtt_broker_connected();
 void mqtt_message_received(char* topic, byte* payload, unsigned int length);
